fix(fastnode): skip trie-state unwind when trie tables are not maintained
In fastnode mode (--engine.skip-state-root-validation) the trie tables
(AccountsTrie/StoragesTrie) are intentionally unmaintained: live-sync
persists empty TrieUpdates and the hashing/Merkle stages are disabled.
unwind_trie_state_from() however still unconditionally computed a trie
revert from changesets and wrote it — walking (near-)empty trie tables on
every deep reorg / unwind, which is both expensive and writes garbage
nodes into tables nothing reads.

- move the fastnode global from reth-engine-primitives to reth-chain-state
  so storage-layer crates can consult it without depending on engine
  crates; re-export from reth_engine_primitives::config so the existing
  activate_fastnode / is_fastnode_active call sites are unchanged
- unwind_trie_state_from: return early before the trie-revert step under
  fastnode; hashed-state and history unwinds still run (hashed tables are
  the canonical state representation under storage v2)

The block-persistence path needs no change: write_trie_updates_sorted is
already guarded by merged_trie.is_empty(), which holds under fastnode.
