+++ crates/engine/primitives/src/config.rs
@@ -1,30 +1,14 @@
 //! Engine tree configuration.
 
 use alloy_eips::merge::EPOCH_SLOTS;
-use core::{
-    sync::atomic::{AtomicBool, Ordering},
-    time::Duration,
-};
+use core::time::Duration;
 
-/// Global flag indicating whether the node is running in fastnode mode
-/// (i.e., `--engine.skip-state-root-validation` is enabled).
-///
-/// When active, trie-dependent RPC methods such as `eth_getProof` and `eth_getAccount`
-/// are unavailable because the hashing stages and state root computation are skipped,
-/// meaning the trie tables in the database are not kept up to date.
-static FASTNODE_ACTIVE: AtomicBool = AtomicBool::new(false);
-
-/// Activate fastnode mode globally.
-///
-/// This should be called once during node startup when `skip_state_root_validation` is enabled.
-pub fn activate_fastnode() {
-    FASTNODE_ACTIVE.store(true, Ordering::SeqCst);
-}
-
-/// Returns `true` if the node is running in fastnode mode.
-pub fn is_fastnode_active() -> bool {
-    FASTNODE_ACTIVE.load(Ordering::SeqCst)
-}
+// The fastnode flag lives in `reth-chain-state` so storage-layer crates (e.g.
+// `reth-provider`, which must skip trie-state unwinds in fastnode mode) can
+// consult it without depending on engine crates. Re-exported here to keep the
+// established `reth_engine_primitives::{activate_fastnode, is_fastnode_active}`
+// paths working.
+pub use reth_chain_state::{activate_fastnode, is_fastnode_active};
 
 /// Triggers persistence when the number of canonical blocks in memory exceeds this threshold.
 pub const DEFAULT_PERSISTENCE_THRESHOLD: u64 = 2;
